refactor: WeatherRepository 파일 디렉토리 이동
