+++ Diary/Infrastructure/Network/NetworkManager/NetworkManager.swift
@@ -7,33 +7,6 @@
 
 import Foundation
 
-final class WeatherRepository {
-    private let networkManager = NetworkManager()
-    private let jsonDecoder = JSONDecoder()
-    
-    func requestAPI (
-        with endpoint: Requestable,
-        completion: @escaping (Result<Weather?, NetworkError>) -> Void
-    ) {
-        networkManager.request(endpoint: endpoint) { [weak self] result in
-            guard let self = self else {
-                return
-            }
-            
-            switch result {
-            case .success(let result):
-                completion(.success(self.decode(data: result)))
-            case .failure(let error):
-                completion(.failure(error))
-            }
-        }
-    }
-    
-    private func decode(data: Data) -> Weather? {
-        return try? jsonDecoder.decode(Weather.self, from: data)
-    }
-}
-
 final class NetworkManager {
     private let session: URLSession
     
@@ -77,17 +50,6 @@ final class NetworkManager {
         case .failure(let error):
             completion(.failure(error))
             return nil
-        }
-    }
-}
-
-fileprivate extension Data {
-    func decode<T: Decodable>(with decoder: JSONDecoder) -> Result<T, Error> {
-        do {
-            let decoded = try decoder.decode(T.self, from: self)
-            return .success(decoded)
-        } catch {
-            return .failure(NetworkError.decodeError)
         }
     }
 }
